move workaround to BaseSchema
