+++ pydtk/db/schemas/dataware-tools.com/v1alpha2/file.py
@@ -1,4 +1,4 @@
-from typing import Optional
+from typing import Optional, Dict, Any, Type
 
 from pydantic import Field, constr, Extra
 
@@ -17,10 +17,13 @@ class File(BaseSchema):
     contents: Optional[dict] = Field(None, description="")
 
     # Allow additional properties
-    class Config:
+    class Config(BaseSchema.Config):
         """Config."""
 
         extra = Extra.allow
-        schema_extra = {
-            "additionalProperties": {}
-        }
+
+        @staticmethod
+        def schema_extra(schema: Dict[str, Any], model: Type['File']):
+            """Extra schema."""
+            BaseSchema.Config.schema_extra(schema, model)
+            schema["additionalProperties"] = {}
